Moved code from UploadView to MappingView
